Move navigation to separate directory
